fix: fix slot decoding and deshreding

+++ src/util/shred.rs
@@ -5,24 +5,27 @@ use solana_ledger::shred::{self, ReedSolomonCache, Shred, ShredType, Shredder};
 
 use crate::{rpc::RpcError, store::shred::ShredStore};
 
-pub fn recover_shreds_and_group_by_completion(
-    shreds_by_fec_set_index: BTreeMap<u32, Vec<Shred>>,
+fn recover_shreds_and_group_by_completion(
+    shreds_by_fec_set_index: BTreeMap<u32, BatchMeta>,
 ) -> Result<impl Iterator<Item = BTreeMap<u32, Shred>>, RpcError> {
     let rs_cache = ReedSolomonCache::default();
     let mut data_shreds_partitioned_by_completion = vec![BTreeMap::<u32, Shred>::new()];
     for shred_batch in shreds_by_fec_set_index.into_values() {
-        let recovered =
-            shred::recover(shred_batch.clone(), &rs_cache).map_err(RpcError::ShredRecovery)?;
         let last = data_shreds_partitioned_by_completion.last_mut().unwrap();
-        for shred in recovered {
-            let shred = shred.map_err(RpcError::ShredRecovery)?;
-            if shred.shred_type() != ShredType::Data {
-                continue;
+        if shred_batch.data_cnt < 32 {
+            let recovered = shred::recover(shred_batch.shreds.clone(), &rs_cache)
+                .map_err(RpcError::ShredRecovery)?;
+            for shred in recovered {
+                let shred = shred.map_err(RpcError::ShredRecovery)?;
+                if shred.shred_type() != ShredType::Data {
+                    continue;
+                }
+                last.insert(shred.index(), shred);
             }
-            last.insert(shred.index(), shred);
         }
         last.extend(
             shred_batch
+                .shreds
                 .into_iter()
                 .filter_map(|s| (s.shred_type() == ShredType::Data).then(|| (s.index(), s))),
         );
@@ -44,18 +47,25 @@ pub fn deshred_to_entries<'a>(
     Ok(deshred_entries)
 }
 
+#[derive(Default)]
+struct BatchMeta {
+    shreds: Vec<Shred>,
+    data_cnt: usize,
+}
+
 pub fn get_slot_entries_from_raw_shreds<S: AsRef<[u8]>>(
     shreds: impl IntoIterator<Item = S>,
 ) -> Result<Vec<Entry>, RpcError> {
-    let mut shreds_for_slot = BTreeMap::<u32, Vec<Shred>>::new();
+    let mut shreds_for_slot = BTreeMap::<u32, BatchMeta>::new();
 
     for shred in shreds {
         let deser = Shred::new_from_serialized_shred(shred.as_ref().to_vec())
             .map_err(RpcError::ShredDeser)?;
-        shreds_for_slot
-            .entry(deser.fec_set_index())
-            .or_default()
-            .push(deser);
+        let meta = shreds_for_slot.entry(deser.fec_set_index()).or_default();
+        if deser.is_data() {
+            meta.data_cnt += 1;
+        }
+        meta.shreds.push(deser);
     }
 
     let mut entries = Vec::new();
@@ -73,22 +83,5 @@ pub fn get_slot_entries_from_store(
 ) -> Result<Vec<Entry>, RpcError> {
     let unsorted_shreds = shred_store.get_slot_shreds(slot)?;
 
-    let mut shreds_for_slot = BTreeMap::<u32, Vec<Shred>>::new();
-
-    for shred in unsorted_shreds {
-        let deser =
-            Shred::new_from_serialized_shred(shred.to_vec()).map_err(RpcError::ShredDeser)?;
-        shreds_for_slot
-            .entry(deser.fec_set_index())
-            .or_default()
-            .push(deser);
-    }
-
-    let mut entries = Vec::new();
-    for data_shreds in recover_shreds_and_group_by_completion(shreds_for_slot)? {
-        let mut deshred_entries = deshred_to_entries(data_shreds.values())?;
-        entries.append(&mut deshred_entries);
-    }
-
-    Ok(entries)
+    get_slot_entries_from_raw_shreds(unsorted_shreds)
 }
